MtJpGraph support added
https://github.com/PHPOffice/PhpSpreadsheet/pull/2979
