+++ phpstan.neon.dist
@@ -11,6 +11,8 @@ parameters:
         - tests/
     excludePaths:
         - src/PhpSpreadsheet/Chart/Renderer/JpGraph.php
+        - src/PhpSpreadsheet/Chart/Renderer/JpGraphRendererBase.php
+        - src/PhpSpreadsheet/Chart/Renderer/MtJpGraphRenderer.php
     parallel:
         processTimeout: 300.0
     checkMissingIterableValueType: false
